requirements.txt
azure-functions
boto3